Updates to repositoryFqn for consistency
Updated the image.repositoryFqn field to be consistent with the
other fields under image.

Previously, repositoryFqn was expected at the same level as the
image: section, now it is expected within the image: section like
other fields (tag, pullPolicy, etc.).

The image tag must now be provided separately from the repositoryFqn.
The repositoryFqn should only be the name of the repository, not the
tag of the specific image.

BRASS-391