Generate valid cert for PingAccess FIPS mode
Added a new privateCert.format field, which can be set to
"pingaccess-fips-pem" to generate a cert that can be used by PingAccess
when running in FIPS mode. The cert is generated by a temporary
PingAccess initContainer, as PingAccess requires a specific format
for certs when in FIPS mode that must be generated from PingAccess
itself.

Leaving the field blank or setting it to any other value will generate
a cert in the same manner as before - adding the key pair to a PKCS12
keystore file.

The externalImage values section has also been updated to expect the
same format for "image:" as the individual products. If no image is
specified for the externalImage, the corresponding defaults from the
product section will be used.

BRASS-433