Fix vault test and improve output of template test script
Fixed the vault template test and improved the template test
script to fail the job when any test fails and to provide
more helpful output.

Also updated the test scripts to truncate the k8s namespace
to prevent hitting the namespace character limit. And updated
a few copyright years.

BRASS-542